pbs: Implement package options
Implement parsing, configuration and save/load of package options.
There are still a couple of issues with this, such as when a
configuration is loaded and is missing a value for a specific option.
Currently such errors are ignored, but eventually maybe a package should
be force-configured in such cases.

Signed-off-by: Thierry Reding <thierry.reding@gmail.com>