devel/llvm: Fix --bindir, --libdir and --includedir
These directories should point to the absolute directories on the host
filesystem. If a build system uses these to find development files it
must prepend the cross-compilation sysroot to make sure the correct
files are found. A preferred way is for build systems to only use the
--cppflags, --cxxflags and --ldflags options to find headers and
libraries, because they will refer to the correct locations within the
sysroot.

Signed-off-by: Thierry Reding <thierry.reding@gmail.com>