build-tools: Export LD_RUN_PATH environment variable
The LD_RUN_PATH environment variable is used by the linker to embed the
path to shared libraries in executables. Specifying this makes sure that
build tools are linked against libraries in the build-tools directory
first. This in turn helps harden the build tools against changes on the
build system.

Signed-off-by: Thierry Reding <thierry.reding@gmail.com>