Remove libs/db package
Ever since Oracle bought this up it's been getting increasingly
difficult to obtain the packages sources for this in an automated way
and to check for new versions. It's only used by the userspace ARP
daemon in net/iproute2, but since we don't use that, let's stop building
it and remove the libs/db dependency and the libs/db package along with
it.

Signed-off-by: Thierry Reding <thierry.reding@gmail.com>